requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.1.0
colorlog==6.6.0
frozenlist==1.3.0
hikari==2.0.0.dev109
hikari-lightbulb==2.2.3
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
yarl==1.7.2
